feat: initial fuse-box bundling support
