+++ index.js
@@ -5,6 +5,6 @@ module.exports = nexe
 
 if (require.main === module) {
   nexe.compile(nexe.argv).catch((e) => {
-    process.stderr.write(e.message, () => process.exit(1))
+    process.stderr.write(e.stack, () => process.exit(1))
   })
 }
