+++ src/bundling/index.ts
@@ -1,0 +1,1 @@
+export * from './fuse-native-module-plugin'
