+++ src/options.ts
@@ -49,6 +49,7 @@ export interface NexeOptions {
    */
   temp: string
   ico?: string
+  sourceMaps?: boolean
   rc: { [key: string]: string }
   /**
    * Causes nexe to remove all temporary files for current configuration
@@ -183,7 +184,7 @@ function extractName(options: NexeOptions) {
 
 function normalizeOptionsAsync(input: Partial<NexeOptions>) {
   if (argv.help || argv._.some((x: string) => x === 'version')) {
-    process.stderr.write(argv.help ? help : '2.0.0-beta.7' + EOL, () => process.exit(0))
+    process.stderr.write(argv.help ? help : '2.0.0-rc.1' + EOL, () => process.exit(0))
   }
 
   const options = Object.assign({}, defaults, input) as NexeOptions
