--- test/embeded-files/hw.txt
@@ -1,1 +1,0 @@
-true
