--- test/gulp-test-170/index.js
@@ -1,2 +1,0 @@
-// it probably executed.
-console.log(true);
