--- test/js-yaml-148/index.js
@@ -1,3 +1,0 @@
-require('js-yaml');
-
-console.log(true);
