--- test/mocha.opts
@@ -1,1 +1,0 @@
---timeout 100000000
